print_app/requirements.txt


pycups
fastapi
uvicorn
